+++ scaffold/simple/trivial.js
@@ -1,0 +1,6 @@
+
+describe('@firstTest@', function () {
+  it('should load a website', async function () {
+    await this.nemo.driver.get(this.nemo.data.baseUrl);
+  });
+});